fix: refactor LeagueDetailPage into smaller components
