+++ src/screens/LeagueDetailPage/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useLeagueDetail';
+
