+++ src/screens/LeagueDetailPage/utils/leagueUtils.ts
@@ -1,0 +1,339 @@
+// Mock data for leagues - using same data as LeaguesPage for now
+export const leagueData = [
+  {
+    id: 1,
+    name: "Elite Womens Volleyball",
+    sport: "Volleyball",
+    format: "6s",
+    day: "Monday",
+    playTimes: ["7:00 PM - 10:00 PM"],
+    location: "Central",
+    specificLocation: "Carleton University",
+    dates: "May 1 - August 15, 2025",
+    skillLevel: "Elite",
+    price: 250,
+    spotsRemaining: 2,
+    isFeatured: true,
+    description: "Our Elite Women's League is designed for players with high-level experience, including current and former university/college players, club players, and those with extensive competitive playing experience. Teams are tiered based on performance to ensure competitive matches each week.",
+    season: "Summer 2025 Season",
+    image: "/womens-elite-card.jpg"
+  },
+  {
+    id: 2,
+    name: "Coed Intermediate Volleyball",
+    sport: "Volleyball",
+    format: "6s",
+    day: "Tuesday",
+    playTimes: ["7:00 PM - 9:00 PM", "9:00 PM - 11:00 PM"],
+    location: "Central",
+    specificLocation: "University of Ottawa",
+    dates: "May 2 - August 16, 2025",
+    skillLevel: "Intermediate",
+    price: 200,
+    spotsRemaining: 8,
+    description: "Our Intermediate Coed 6's League is perfect for players with at least moderate volleyball experience. Teams should have consistent serving, passing, and attacking abilities. Play begins with proper 3-hit sequences and basic offensive strategies.",
+    season: "Summer 2025 Season",
+    image: "/571North-CR3_0335-Indoor-VB-Header-Featured.jpg"
+  },
+  {
+    id: 3,
+    name: "Advanced Badminton",
+    sport: "Badminton",
+    format: "Singles",
+    day: "Wednesday",
+    playTimes: ["6:30 PM - 9:30 PM"],
+    location: "East End",
+    specificLocation: "Rideau High School",
+    dates: "May 3 - August 17, 2025",
+    skillLevel: "Advanced",
+    price: 180,
+    spotsRemaining: 0,
+    isFeatured: true,
+    description: "Our Advanced Badminton Singles League is for experienced players with refined technique and strategic understanding of the game. Players should have strong shot variety, court movement, and competitive experience.",
+    season: "Summer 2025 Season",
+    image: "/badminton-card.png"
+  },
+  {
+    id: 4,
+    name: "Coed Competitive Volleyball",
+    sport: "Volleyball",
+    format: "4s",
+    day: "Thursday",
+    playTimes: ["7:00 PM - 9:00 PM", "9:00 PM - 11:00 PM"],
+    location: "Central",
+    specificLocation: "Glebe Collegiate",
+    dates: "May 4 - August 18, 2025",
+    skillLevel: "Competitive",
+    price: 220,
+    spotsRemaining: 4,
+    isFeatured: true,
+    description: "Our Competitive Coed 4's League offers a challenging environment for experienced players. Teams should have strong positional play, offensive and defensive systems, and the ability to run complex plays. This league features tiered play to ensure competitive matches.",
+    season: "Summer 2025 Season",
+    image: "/571North-CR3_0335-Indoor-VB-Header-Featured.jpg"
+  },
+  {
+    id: 5,
+    name: "Intermediate Pickleball",
+    sport: "Pickleball",
+    format: "Doubles",
+    day: "Friday",
+    playTimes: ["6:00 PM - 9:00 PM"],
+    location: "Central",
+    specificLocation: "Carleton University",
+    dates: "May 5 - August 19, 2025",
+    skillLevel: "Intermediate",
+    price: 150,
+    spotsRemaining: 12,
+    description: "Our Intermediate Pickleball Doubles League is for players who understand the rules and have developed consistent strokes. Players should be able to sustain rallies and are beginning to use strategies and shot placement.",
+    season: "Summer 2025 Season",
+    image: "/pickleball-card.jpg"
+  },
+  {
+    id: 6,
+    name: "Mens Advanced Volleyball",
+    sport: "Volleyball",
+    format: "6s",
+    day: "Saturday",
+    playTimes: ["10:00 AM - 1:00 PM", "2:00 PM - 5:00 PM"],
+    location: "West End",
+    specificLocation: "Nepean Sportsplex",
+    dates: "May 6 - August 20, 2025",
+    skillLevel: "Advanced",
+    price: 230,
+    spotsRemaining: 6,
+    description: "Our Men's Advanced 6's League is designed for players with significant volleyball experience. Teams should have strong serving, passing, and hitting abilities, with play featuring consistent 3-hit sequences and organized offensive strategies. This league uses a tiered system based on team performance.",
+    season: "Summer 2025 Season",
+    image: "/indoor-coed.jpg"
+  },
+  {
+    id: 7,
+    name: "Competitive Basketball",
+    sport: "Basketball",
+    format: "5s",
+    day: "Sunday",
+    playTimes: ["12:00 PM - 3:00 PM"],
+    location: "East End",
+    specificLocation: "Orleans Recreation Complex",
+    dates: "May 7 - August 21, 2025",
+    skillLevel: "Competitive",
+    price: 190,
+    spotsRemaining: 3,
+    isFeatured: true,
+    description: "Our Competitive Basketball League is for experienced players who understand team concepts and have refined individual skills. Games feature organized offensive and defensive strategies, with an emphasis on teamwork and basketball IQ.",
+    season: "Summer 2025 Season",
+    image: "/indoor-coed.jpg"
+  },
+  {
+    id: 8,
+    name: "Coed Intermediate Badminton",
+    sport: "Badminton",
+    format: "Doubles",
+    day: "Monday",
+    playTimes: ["7:30 PM - 10:30 PM"],
+    location: "Central",
+    specificLocation: "Glebe Collegiate",
+    dates: "May 8 - August 22, 2025",
+    skillLevel: "Intermediate",
+    price: 175,
+    spotsRemaining: 8,
+    description: "Our Intermediate Coed Badminton Doubles League is perfect for players who have developed consistent strokes and are beginning to implement strategy. Players should understand proper positioning, shot selection, and have some competitive experience.",
+    season: "Summer 2025 Season",
+    image: "/badminton-card.png"
+  },
+  {
+    id: 9,
+    name: "Elite Coed Volleyball",
+    sport: "Volleyball",
+    format: "4s",
+    day: "Tuesday & Thursday",
+    playTimes: ["7:00 PM - 10:00 PM", "8:00 PM - 11:00 PM"],
+    location: "West End",
+    specificLocation: "Kanata Recreation Complex",
+    dates: "May 9 - August 23, 2025",
+    skillLevel: "Elite",
+    price: 275,
+    spotsRemaining: 1,
+    isFeatured: true,
+    description: "Our Elite Coed 4's League is our highest level of play, designed for advanced players with extensive competitive experience. Play features complex offensive systems, strategic defenses, and high-level execution. This league utilizes a strict tiering system to ensure balanced competition.",
+    season: "Summer 2025 Season",
+    image: "/571North-CR3_0335-Indoor-VB-Header-Featured.jpg"
+  }
+];
+
+// Mock data for standings
+export const mockStandings = [
+  { id: 1, team: "Setters of Catan", wins: 18, losses: 2, points: 36, pointsAgainst: 10, differentials: "+26" },
+  { id: 2, team: "Block Party", wins: 16, losses: 4, points: 32, pointsAgainst: 12, differentials: "+20" },
+  { id: 3, team: "Dig It", wins: 15, losses: 5, points: 30, pointsAgainst: 14, differentials: "+16" },
+  { id: 4, team: "Net Gains", wins: 14, losses: 6, points: 28, pointsAgainst: 18, differentials: "+10" },
+  { id: 5, team: "Serve-ivors", wins: 13, losses: 7, points: 26, pointsAgainst: 18, differentials: "+8" },
+  { id: 6, team: "Spiked Punch", wins: 12, losses: 8, points: 24, pointsAgainst: 20, differentials: "+4" },
+  { id: 7, team: "Setting Ducks", wins: 11, losses: 9, points: 22, pointsAgainst: 22, differentials: "0" },
+  { id: 8, team: "Block and Roll", wins: 10, losses: 10, points: 20, pointsAgainst: 22, differentials: "-2" },
+  { id: 9, team: "Ace of Base", wins: 9, losses: 11, points: 18, pointsAgainst: 24, differentials: "-6" },
+  { id: 10, team: "Sets on the Beach", wins: 8, losses: 12, points: 16, pointsAgainst: 26, differentials: "-10" },
+  { id: 11, team: "Hitting on All Sixes", wins: 7, losses: 13, points: 14, pointsAgainst: 28, differentials: "-14" },
+  { id: 12, team: "Set to Kill", wins: 6, losses: 14, points: 12, pointsAgainst: 30, differentials: "-18" },
+  { id: 13, team: "Bump Set Psycho", wins: 5, losses: 15, points: 10, pointsAgainst: 32, differentials: "-22" },
+  { id: 14, team: "Net Assets", wins: 4, losses: 16, points: 8, pointsAgainst: 34, differentials: "-26" },
+  { id: 15, team: "Notorious DIG", wins: 3, losses: 17, points: 6, pointsAgainst: 36, differentials: "-30" },
+  { id: 16, team: "Volley Llamas", wins: 2, losses: 18, points: 4, pointsAgainst: 38, differentials: "-34" },
+  { id: 17, team: "The Ace Holes", wins: 1, losses: 19, points: 2, pointsAgainst: 40, differentials: "-38" },
+  { id: 18, team: "Served Cold", wins: 0, losses: 20, points: 0, pointsAgainst: 42, differentials: "-42" },
+  { id: 19, team: "Spike Force", wins: 0, losses: 21, points: 0, pointsAgainst: 44, differentials: "-44" },
+  { id: 20, team: "Setting Standards", wins: 0, losses: 22, points: 0, pointsAgainst: 46, differentials: "-46" },
+  { id: 21, team: "Dig or Die", wins: 0, losses: 23, points: 0, pointsAgainst: 48, differentials: "-48" },
+  { id: 22, team: "Blocking Legends", wins: 0, losses: 24, points: 0, pointsAgainst: 50, differentials: "-50" },
+  { id: 23, team: "Serve's Up", wins: 0, losses: 25, points: 0, pointsAgainst: 52, differentials: "-52" },
+  { id: 24, team: "The Dig Dudes", wins: 0, losses: 26, points: 0, pointsAgainst: 54, differentials: "-54" },
+  { id: 25, team: "Bump Set Spike", wins: 0, losses: 27, points: 0, pointsAgainst: 56, differentials: "-56" },
+  { id: 26, team: "Net Navigators", wins: 0, losses: 28, points: 0, pointsAgainst: 58, differentials: "-58" },
+  { id: 27, team: "Passing Passion", wins: 0, losses: 29, points: 0, pointsAgainst: 60, differentials: "-60" },
+  { id: 28, team: "Rotation Nation", wins: 0, losses: 30, points: 0, pointsAgainst: 62, differentials: "-62" },
+  { id: 29, team: "Court Crushers", wins: 0, losses: 31, points: 0, pointsAgainst: 64, differentials: "-64" },
+  { id: 30, team: "Libero Legends", wins: 0, losses: 32, points: 0, pointsAgainst: 66, differentials: "-66" },
+  { id: 31, team: "Bump Brigade", wins: 0, losses: 33, points: 0, pointsAgainst: 68, differentials: "-68" },
+  { id: 32, team: "The Backrow Bombers", wins: 0, losses: 34, points: 0, pointsAgainst: 70, differentials: "-70" },
+  { id: 33, team: "Volley Vikings", wins: 0, losses: 35, points: 0, pointsAgainst: 72, differentials: "-72" },
+  { id: 34, team: "Setters Paradise", wins: 0, losses: 36, points: 0, pointsAgainst: 74, differentials: "-74" },
+  { id: 35, team: "Ace Attackers", wins: 0, losses: 37, points: 0, pointsAgainst: 76, differentials: "-76" },
+  { id: 36, team: "The Last Set", wins: 0, losses: 38, points: 0, pointsAgainst: 78, differentials: "-78" }
+];
+
+// Get background color based on skill level
+export const getSkillLevelColor = (level: string) => {
+  switch (level) {
+    case 'Elite':
+      return 'bg-purple-100 text-purple-800';
+    case 'Competitive':
+      return 'bg-blue-100 text-blue-800';
+    case 'Advanced':
+      return 'bg-indigo-100 text-indigo-800';
+    case 'Intermediate':
+      return 'bg-teal-100 text-teal-800';
+    default:
+      return 'bg-gray-100 text-gray-800';
+  }
+};
+
+// Function to get sport icon based on sport type
+export const getSportIcon = (sport: string) => {
+  switch (sport) {
+    case 'Volleyball':
+      return "/Volleyball.png";
+    case 'Badminton':
+      return "/Badminton.png";
+    case 'Basketball':
+      return "/Basketball.png";
+    case 'Pickleball':
+      return "/Pickleball.png";
+    default:
+      return "";
+  }
+};
+
+// Function to get spots badge color
+export const getSpotsBadgeColor = (spots: number) => {
+  if (spots === 0) return "bg-red-100 text-red-800";
+  if (spots <= 3) return "bg-orange-100 text-orange-800";
+  return "bg-green-100 text-green-800";
+};
+
+// Function to get spots text
+export const getSpotsText = (spots: number) => {
+  if (spots === 0) return "Full";
+  if (spots === 1) return "1 spot left";
+  return `${spots} spots left`;
+};
+
+// Get team name from position
+export const getTeamNameFromPosition = (tier: any, position: string) => {
+  return tier.teams[position]?.name || "";
+};
+
+// Types for schedule data
+export interface Team {
+  name: string;
+  ranking: number;
+}
+
+export interface Tier {
+  tierNumber: number;
+  location: string;
+  time: string;
+  court: string;
+  teams: Record<string, Team | null>;
+  courts: Record<string, string>;
+}
+
+export interface Schedule {
+  week: number;
+  date: string;
+  tiers: Tier[];
+}
+
+// Function to generate tiered schedule based on teams
+export const generateTieredSchedule = (): Schedule => {
+  const teamsPerTier = 3;
+  const totalTiers = Math.ceil(mockStandings.length / teamsPerTier);
+  const tiers: Tier[] = [];
+  
+  // Locations to cycle through
+  const locations = [
+    "Carleton University", 
+    "University of Ottawa", 
+    "Glebe Collegiate", 
+    "Nepean Sportsplex", 
+    "Orleans Recreation Complex"
+  ];
+  
+  // Time slots to cycle through
+  const timeSlots = [
+    "7:00 PM - 8:30 PM",
+    "8:30 PM - 10:00 PM",
+    "6:00 PM - 7:30 PM",
+    "7:30 PM - 9:00 PM"
+  ];
+  
+  // Generate each tier
+  for (let i = 0; i < totalTiers; i++) {
+    const startIndex = i * teamsPerTier;
+    const tierTeams = mockStandings.slice(startIndex, startIndex + teamsPerTier);
+    
+    // If we don't have exactly 3 teams, fill with placeholders
+    const teams: Record<string, { name: string; ranking: number } | null> = { A: null, B: null, C: null };
+    const courts = { A: `Court ${(i % 3) + 1}`, B: `Court ${(i % 3) + 1}`, C: `Court ${(i % 3) + 1}` };
+    
+    // Assign available teams with their rankings
+    tierTeams.forEach((team, index) => {
+      const position = String.fromCharCode(65 + index); // A, B, C
+      teams[position] = {
+        name: team.team,
+        ranking: startIndex + index + 1 // Ranking based on position in mockStandings
+      };
+    });
+    
+    // Only create a tier if we have at least one team
+    if (teams.A) {
+      // Add tier to the schedule
+      tiers.push({
+        tierNumber: i + 1,
+        location: locations[i % locations.length],
+        time: timeSlots[i % timeSlots.length],
+        court: `Court ${(i % 3) + 1}`,
+        teams: teams,
+        courts: courts
+      });
+    }
+  }
+  
+  return {
+    week: 1,
+    date: "June 5, 2025",
+    tiers: tiers
+  };
+};
+
+// Generate the schedule
+export const mockSchedule = [generateTieredSchedule()];
+
